Add security endpoint with access Token

+++ part3/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -17,10 +18,30 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         try:
+            current_user_id = get_jwt_identity()
+
+            place_id = api.payload["place_id"]
+
+            # récupérer le place
+            place = facade.get_place(place_id)
+
+            # vérifier si l'utilisateur est le propriétaire
+            if place.owner.id == current_user_id:
+                return {"error": "You cannot review your own place."}, 400
+
+            reviews = facade.get_reviews_by_place(place_id)
+
+            # vérifier si l'utilisateur a déjà fait une review
+            for review in reviews:
+                if review.user.id == current_user_id:
+                    return {"error": "You have already reviewed this place."}, 400
+
             review = facade.create_review(api.payload)
+
             return {
                 'id': review.id,
                 'text': review.text,
@@ -65,25 +86,38 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Place not found'}, 404
 
+        if review.user.id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
         try:
             facade.update_review(review_id, api.payload)
-            return {'message': 'Review deleted successfully'}, 200
+            return {'message': 'Review updated successfully'}, 200
         except ValueError as e:
             return {'error': str(e)}, 400
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user.id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
         facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
